requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.0
redis==4.5.5
django-celery-results==2.5.0



 